+++ core/src/main/java/com/graphhopper/routing/util/parsers/AbstractAccessParser.java
@@ -33,7 +33,7 @@ public abstract class AbstractAccessParser implements TagParser {
     protected final List<String> restrictionKeys;
     protected final Set<String> restrictedValues = new HashSet<>(5);
 
-    protected final Set<String> intendedValues = new HashSet<>(INTENDED); // possible to add "private" later
+    protected final Set<String> allowedValues = new HashSet<>(INTENDED);
     // http://wiki.openstreetmap.org/wiki/Mapfeatures#Barrier
     protected final Set<String> barriers = new HashSet<>(5);
     protected final BooleanEncodedValue accessEnc;
@@ -43,11 +43,16 @@ public abstract class AbstractAccessParser implements TagParser {
         this.accessEnc = accessEnc;
         this.restrictionKeys = restrictionKeys;
 
+        allowedValues.add("destination");
+
         restrictedValues.add("no");
         restrictedValues.add("restricted");
         restrictedValues.add("military");
         restrictedValues.add("emergency");
+        restrictedValues.add("unknown");
+
         restrictedValues.add("private");
+        restrictedValues.add("service");
         restrictedValues.add("permit");
     }
 
@@ -61,12 +66,11 @@ public abstract class AbstractAccessParser implements TagParser {
 
     protected void blockPrivate(boolean blockPrivate) {
         if (!blockPrivate) {
-            if (!restrictedValues.remove("private"))
-                throw new IllegalStateException("no 'private' found in restrictedValues");
-            if (!restrictedValues.remove("permit"))
-                throw new IllegalStateException("no 'permit' found in restrictedValues");
-            intendedValues.add("private");
-            intendedValues.add("permit");
+            if (!restrictedValues.remove("private") || !restrictedValues.remove("permit") || !restrictedValues.remove("service"))
+                throw new IllegalStateException("no 'private', 'permit' or 'service' value found in restrictedValues");
+            allowedValues.add("private");
+            allowedValues.add("permit");
+            allowedValues.add("service");
         }
     }
 
@@ -97,9 +101,9 @@ public abstract class AbstractAccessParser implements TagParser {
 
         if (restrictedValues.contains(firstValue))
             return true;
-        else if (node.hasTag("locked", "yes") && !intendedValues.contains(firstValue))
+        else if (node.hasTag("locked", "yes") && !allowedValues.contains(firstValue))
             return true;
-        else if (intendedValues.contains(firstValue))
+        else if (allowedValues.contains(firstValue))
             return false;
         else if (node.hasTag("barrier", barriers))
             return true;