+++ core/src/main/java/com/graphhopper/routing/util/parsers/BikeCommonPriorityParser.java
@@ -10,15 +10,16 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static com.graphhopper.routing.ev.RouteNetwork.*;
 import static com.graphhopper.routing.util.PriorityCode.*;
 import static com.graphhopper.routing.util.parsers.AbstractAccessParser.INTENDED;
 
 public abstract class BikeCommonPriorityParser implements TagParser {
-    // Bicycle tracks subject to compulsory use in Germany and Poland (https://wiki.openstreetmap.org/wiki/DE:Key:cycleway)
+    private static final Set<String> CYCLEWAY_KEYS = Set.of("cycleway", "cycleway:left", "cycleway:both", "cycleway:right");
+
+    // rare use case when a bicycle lane has access tag
     private static final List<String> CYCLEWAY_BICYCLE_KEYS = List.of("cycleway:bicycle", "cycleway:both:bicycle", "cycleway:left:bicycle", "cycleway:right:bicycle");
 
-    // Pushing section highways are parts where you need to get off your bike and push it (German: Schiebestrecke)
+    // pushing section highways are parts where you need to get off your bike and push it
     protected final HashSet<String> pushingSectionsHighways = new HashSet<>();
     protected final Set<String> preferHighwayTags = new HashSet<>();
     protected final Map<String, PriorityCode> avoidHighwayTags = new HashMap<>();
@@ -28,20 +29,15 @@ public abstract class BikeCommonPriorityParser implements TagParser {
     protected final DecimalEncodedValue priorityEnc;
     // Car speed limit which switches the preference from UNCHANGED to AVOID_IF_POSSIBLE
     int avoidSpeedLimit;
-    EnumEncodedValue<RouteNetwork> bikeRouteEnc;
-    Map<RouteNetwork, Integer> routeMap = new HashMap<>();
     protected final Set<String> goodSurface = Set.of("paved", "asphalt", "concrete");
 
     // This is the specific bicycle class
     private String classBicycleKey;
 
-    protected BikeCommonPriorityParser(DecimalEncodedValue priorityEnc, DecimalEncodedValue avgSpeedEnc,
-                                       EnumEncodedValue<RouteNetwork> bikeRouteEnc) {
-        this.bikeRouteEnc = bikeRouteEnc;
+    protected BikeCommonPriorityParser(DecimalEncodedValue priorityEnc, DecimalEncodedValue avgSpeedEnc) {
         this.priorityEnc = priorityEnc;
         this.avgSpeedEnc = avgSpeedEnc;
 
-        // duplicate code as also in BikeCommonAverageSpeedParser
         addPushingSection("footway");
         addPushingSection("pedestrian");
         addPushingSection("steps");
@@ -66,16 +62,11 @@ public abstract class BikeCommonPriorityParser implements TagParser {
         avoidHighwayTags.put("motorway_link", REACH_DESTINATION);
         avoidHighwayTags.put("trunk", REACH_DESTINATION);
         avoidHighwayTags.put("trunk_link", REACH_DESTINATION);
-        avoidHighwayTags.put("primary", AVOID);
-        avoidHighwayTags.put("primary_link", AVOID);
+        avoidHighwayTags.put("primary", BAD);
+        avoidHighwayTags.put("primary_link", BAD);
         avoidHighwayTags.put("secondary", AVOID);
         avoidHighwayTags.put("secondary_link", AVOID);
         avoidHighwayTags.put("bridleway", AVOID);
-
-        routeMap.put(INTERNATIONAL, BEST.getValue());
-        routeMap.put(NATIONAL, BEST.getValue());
-        routeMap.put(REGIONAL, VERY_NICE.getValue());
-        routeMap.put(LOCAL, VERY_NICE.getValue());
 
         avoidSpeedLimit = 71;
     }
@@ -83,63 +74,40 @@ public abstract class BikeCommonPriorityParser implements TagParser {
     @Override
     public void handleWayTags(int edgeId, EdgeIntAccess edgeIntAccess, ReaderWay way, IntsRef relationFlags) {
         String highwayValue = way.getTag("highway");
-        Integer priorityFromRelation = routeMap.get(bikeRouteEnc.getEnum(false, edgeId, edgeIntAccess));
-        RouteNetwork bikeNetork = this.bikeRouteEnc.getEnum(false, edgeId, edgeIntAccess);
-        if (highwayValue == null) {
+        TreeMap<Double, PriorityCode> weightToPrioMap = new TreeMap<>();
+        if (highwayValue != null) {
+            weightToPrioMap.put(0d, UNCHANGED);
+        } else {
             if (FerrySpeedCalculator.isFerry(way)) {
-                priorityFromRelation = SLIGHT_AVOID.getValue();
+                weightToPrioMap.put(110d, SLIGHT_AVOID);
             } else {
                 return;
             }
         }
 
-        double maxSpeed = Math.max(avgSpeedEnc.getDecimal(false, edgeId, edgeIntAccess), avgSpeedEnc.getDecimal(true, edgeId, edgeIntAccess));
-        priorityEnc.setDecimal(false, edgeId, edgeIntAccess, PriorityCode.getValue(handlePriority(way, maxSpeed, priorityFromRelation, bikeNetork)));
-    }
+        double maxSpeed = Math.max(avgSpeedEnc.getDecimal(false, edgeId, edgeIntAccess),
+                avgSpeedEnc.getDecimal(true, edgeId, edgeIntAccess));
+        collect(way, maxSpeed, isBikeDesignated(way), weightToPrioMap);
 
-    /**
-     * In this method we prefer cycleways or roads with designated bike access and avoid big roads
-     * or roads with trams or pedestrian.
-     *
-     * @return new priority based on priorityFromRelation and on the tags in ReaderWay.
-     */
-    int handlePriority(ReaderWay way, double wayTypeSpeed, Integer priorityFromRelation, RouteNetwork bikeNetork) {
-        TreeMap<Double, PriorityCode> weightToPrioMap = new TreeMap<>();
-        if (priorityFromRelation == null)
-            weightToPrioMap.put(0d, UNCHANGED);
-        else
-            weightToPrioMap.put(110d, PriorityCode.valueOf(priorityFromRelation));
-
-        collect(way, wayTypeSpeed, weightToPrioMap, bikeNetork);
-
-        // pick priority with biggest order value
-        return weightToPrioMap.lastEntry().getValue().getValue();
+        // pick priority with the biggest order value
+        double prio = PriorityCode.getValue(weightToPrioMap.lastEntry().getValue().getValue());
+        priorityEnc.setDecimal(false, edgeId, edgeIntAccess, prio);
     }
 
     // Conversion of class value to priority. See http://wiki.openstreetmap.org/wiki/Class:bicycle
     private PriorityCode convertClassValueToPriority(String tagvalue) {
-        int classvalue;
         try {
-            classvalue = Integer.parseInt(tagvalue);
+            return switch (Integer.parseInt(tagvalue)) {
+                case 3 -> BEST;
+                case 2 -> VERY_NICE;
+                case 1 -> PREFER;
+                case -1 -> AVOID;
+                case -2 -> BAD;
+                case -3 -> REACH_DESTINATION;
+                default -> UNCHANGED;
+            };
         } catch (NumberFormatException e) {
             return UNCHANGED;
-        }
-
-        switch (classvalue) {
-            case 3:
-                return BEST;
-            case 2:
-                return VERY_NICE;
-            case 1:
-                return PREFER;
-            case -1:
-                return SLIGHT_AVOID;
-            case -2:
-                return AVOID;
-            case -3:
-                return AVOID_MORE;
-            default:
-                return UNCHANGED;
         }
     }
 
@@ -147,11 +115,10 @@ public abstract class BikeCommonPriorityParser implements TagParser {
      * @param weightToPrioMap associate a weight with every priority. This sorted map allows
      *                        subclasses to 'insert' more important priorities as well as overwrite determined priorities.
      */
-    void collect(ReaderWay way, double wayTypeSpeed, TreeMap<Double, PriorityCode> weightToPrioMap, RouteNetwork bikeNetork) {
+    void collect(ReaderWay way, double wayTypeSpeed, boolean bikeDesignated, TreeMap<Double, PriorityCode> weightToPrioMap) {
         String highway = way.getTag("highway");
-
-        if (isDesignated(way)) {
-            boolean isGoodSurface = way.getTag("tracktype", "").equals("grade1") || goodSurface.contains(way.getTag("surface",""));
+        if (bikeDesignated) {
+            boolean isGoodSurface = way.getTag("tracktype", "").equals("grade1") || goodSurface.contains(way.getTag("surface", ""));
             if ("path".equals(highway) || "track".equals(highway) && isGoodSurface)
                 weightToPrioMap.put(100d, VERY_NICE);
             else
@@ -166,7 +133,7 @@ public abstract class BikeCommonPriorityParser implements TagParser {
         }
 
         double maxSpeed = Math.max(OSMMaxSpeedParser.parseMaxSpeed(way, false), OSMMaxSpeedParser.parseMaxSpeed(way, true));
-        if (preferHighwayTags.contains(highway) || (maxSpeed != MaxSpeed.MAXSPEED_MISSING && maxSpeed <= 30)) {
+        if (preferHighwayTags.contains(highway) || maxSpeed <= 30) {
             if (maxSpeed == MaxSpeed.MAXSPEED_MISSING || maxSpeed < avoidSpeedLimit) {
                 weightToPrioMap.put(40d, PREFER);
                 if (way.hasTag("tunnel", INTENDED))
@@ -175,15 +142,7 @@ public abstract class BikeCommonPriorityParser implements TagParser {
         } else if (avoidHighwayTags.containsKey(highway)
                 || (maxSpeed != MaxSpeed.MAXSPEED_MISSING && maxSpeed >= avoidSpeedLimit && !"track".equals(highway))) {
             PriorityCode priorityCode = avoidHighwayTags.get(highway);
-            
-
-            if (bikeNetork != RouteNetwork.MISSING) {
-                // If there's a bike network on an avoided highway, give it a better priority than REACH_DESTINATION
-                weightToPrioMap.put(50d, AVOID);
-            } else {
-                weightToPrioMap.put(50d, priorityCode == null ? AVOID : priorityCode);
-            }
-            
+            weightToPrioMap.put(50d, priorityCode == null ? AVOID : priorityCode);
             if (way.hasTag("tunnel", INTENDED)) {
                 PriorityCode worse = priorityCode == null ? BAD : priorityCode.worse().worse();
                 weightToPrioMap.put(50d, worse == EXCLUDE ? REACH_DESTINATION : worse);
@@ -205,7 +164,7 @@ public abstract class BikeCommonPriorityParser implements TagParser {
                 pushingSectionPrio = BAD;
             else if (way.hasTag("bicycle", "yes") || way.hasTag("bicycle", "permissive"))
                 pushingSectionPrio = PREFER;
-            else if (isDesignated(way))
+            else if (bikeDesignated)
                 pushingSectionPrio = VERY_NICE;
 
             if (way.hasTag("foot", "yes") && !way.hasTag("segregated", "yes"))
@@ -217,17 +176,16 @@ public abstract class BikeCommonPriorityParser implements TagParser {
         if (way.hasTag("railway", "tram"))
             weightToPrioMap.put(50d, AVOID_MORE);
 
-        if (way.hasTag("lcn", "yes"))
-            weightToPrioMap.put(100d, VERY_NICE);
-
         String classBicycleValue = way.getTag(classBicycleKey);
+        if (classBicycleValue == null) classBicycleValue = way.getTag("class:bicycle");
+
+        // We assume that humans are better in classifying preferences compared to our algorithm above
         if (classBicycleValue != null) {
-            // We assume that humans are better in classifying preferences compared to our algorithm above -> weight = 100
-            weightToPrioMap.put(100d, convertClassValueToPriority(classBicycleValue));
-        } else {
-            String classBicycle = way.getTag("class:bicycle");
-            if (classBicycle != null)
-                weightToPrioMap.put(100d, convertClassValueToPriority(classBicycle));
+            PriorityCode prio = convertClassValueToPriority(classBicycleValue);
+            // do not overwrite if e.g. designated
+            weightToPrioMap.compute(100d, (key, existing) ->
+                    existing == null || existing.getValue() < prio.getValue() ? prio : existing
+            );
         }
 
         // Increase the priority for scenic routes or in case that maxspeed limits our average speed as compensation. See #630
@@ -238,12 +196,16 @@ public abstract class BikeCommonPriorityParser implements TagParser {
         }
     }
 
-    boolean isDesignated(ReaderWay way) {
-        return way.hasTag("bicycle", "designated") || way.hasTag(CYCLEWAY_BICYCLE_KEYS, "designated")
-                || way.hasTag("bicycle_road", "yes") || way.hasTag("cyclestreet", "yes") || way.hasTag("bicycle", "official");
+    static boolean isBikeDesignated(ReaderWay way) {
+        return way.hasTag("bicycle", "designated")
+                || way.hasTag("bicycle", "official")
+                || way.hasTag("segregated", "yes")
+                || way.hasTag("bicycle_road", "yes")
+                || way.hasTag("cyclestreet", "yes")
+                || CYCLEWAY_KEYS.stream().anyMatch(k -> way.getTag(k, "").equals("track"))
+                || way.hasTag(CYCLEWAY_BICYCLE_KEYS, "designated");
     }
 
-    // TODO duplicated in average speed
     void addPushingSection(String highway) {
         pushingSectionsHighways.add(highway);
     }