+++ core/src/test/java/com/graphhopper/GraphHopperTest.java
@@ -104,8 +104,8 @@ public class GraphHopperTest {
             ASTAR + ",false,361",
             DIJKSTRA_BI + ",false,340",
             ASTAR_BI + ",false,192",
-            ASTAR_BI + ",true,46",
-            DIJKSTRA_BI + ",true,51"
+            DIJKSTRA_BI + ",true,45",
+            ASTAR_BI + ",true,43",
     })
     public void testMonacoDifferentAlgorithms(String algo, boolean withCH, int expectedVisitedNodes) {
         GraphHopper hopper = new GraphHopper().
@@ -157,7 +157,7 @@ public class GraphHopperTest {
                 setAlgorithm(ASTAR).setProfile(profile));
 
         // identify the number of counts to compare with CH foot route
-        assertEquals(1033, rsp.getHints().getLong("visited_nodes.sum", 0));
+        assertEquals(1040, rsp.getHints().getLong("visited_nodes.sum", 0));
 
         ResponsePath res = rsp.getBest();
         assertEquals(3536, res.getDistance(), 1);
@@ -213,20 +213,20 @@ public class GraphHopperTest {
         // no simplification
         hopper.getRouterConfig().setSimplifyResponse(false);
         GHResponse routeRsp = hopper.route(request);
-        assertEquals(8, routeRsp.getBest().getInstructions().size());
-        assertEquals(50, routeRsp.getBest().getPoints().size());
+        assertEquals(10, routeRsp.getBest().getInstructions().size());
+        assertEquals(52, routeRsp.getBest().getPoints().size());
 
         // with simplification
         hopper.getRouterConfig().setSimplifyResponse(true);
         routeRsp = hopper.route(request);
-        assertEquals(8, routeRsp.getBest().getInstructions().size());
-        assertEquals(46, routeRsp.getBest().getPoints().size());
+        assertEquals(10, routeRsp.getBest().getInstructions().size());
+        assertEquals(50, routeRsp.getBest().getPoints().size());
 
         // no instructions
         request.getHints().putObject("instructions", false);
         routeRsp = hopper.route(request);
         // the path is still simplified
-        assertEquals(46, routeRsp.getBest().getPoints().size());
+        assertEquals(50, routeRsp.getBest().getPoints().size());
     }
 
     @Test
@@ -445,29 +445,27 @@ public class GraphHopperTest {
 
     @Test
     public void testAlternativeRoutesBike() {
-        final String profile = "profile";
-
         GraphHopper hopper = new GraphHopper().
                 setGraphHopperLocation(GH_LOCATION).
                 setOSMFile(BAYREUTH).
                 setEncodedValuesString("car_access, car_average_speed, bike_access, bike_priority, bike_average_speed").
-                setProfiles(TestProfiles.accessSpeedAndPriority(profile, "bike"));
+                setProfiles(TestProfiles.accessSpeedAndPriority("bike", "bike"));
         hopper.importOrLoad();
 
-        GHRequest req = new GHRequest(50.028917, 11.496506, 49.982089,11.599224).
-                setAlgorithm(ALT_ROUTE).setProfile(profile);
+        GHRequest req = new GHRequest(50.028917, 11.496506, 49.981979, 11.591156).
+                setAlgorithm(ALT_ROUTE).setProfile("bike");
 
         req.putHint("alternative_route.max_paths", 3);
         GHResponse rsp = hopper.route(req);
         assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
 
         assertEquals(3, rsp.getAll().size());
+        // via obergräfenthal
+        assertEquals(2651, rsp.getAll().get(0).getTime() / 1000);
         // via ramsenthal
-        assertEquals(2636, rsp.getAll().get(0).getTime() / 1000);
-        // via eselslohe
-        assertEquals(2783, rsp.getAll().get(1).getTime() / 1000);
+        assertEquals(2771, rsp.getAll().get(1).getTime() / 1000);
         // via unterwaiz
-        assertEquals(2985, rsp.getAll().get(2).getTime() / 1000);
+        assertEquals(2850, rsp.getAll().get(2).getTime() / 1000);
     }
 
     @Test
@@ -1282,11 +1280,11 @@ public class GraphHopperTest {
                 setProfile(bikeProfile));
         assertFalse(rsp.hasErrors());
         ResponsePath res = rsp.getBest();
-        assertEquals(6932.2, res.getDistance(), .1);
-        assertEquals(117, res.getPoints().size());
+        assertEquals(7007.7, res.getDistance(), .1);
+        assertEquals(136, res.getPoints().size());
 
         InstructionList il = res.getInstructions();
-        assertEquals(19, il.size());
+        assertEquals(25, il.size());
 
         assertEquals("continue onto Obere Landstraße", il.get(0).getTurnDescription(tr));
         assertEquals(69.28, (Double) il.get(0).getExtraInfoJSON().get("heading"), .01);
@@ -1297,12 +1295,9 @@ public class GraphHopperTest {
         assertEquals("keep left onto Hoher Markt", il.get(4).getTurnDescription(tr));
         assertEquals("turn right onto Wegscheid", il.get(6).getTurnDescription(tr));
         assertEquals("continue onto Wegscheid", il.get(7).getTurnDescription(tr));
-        assertEquals("turn right onto Ringstraße", il.get(8).getTurnDescription(tr));
-        assertEquals("keep left onto Eyblparkstraße", il.get(9).getTurnDescription(tr));
-        assertEquals("keep left onto Austraße", il.get(10).getTurnDescription(tr));
-        assertEquals("keep left onto Rechte Kremszeile", il.get(11).getTurnDescription(tr));
+        assertEquals("at roundabout, take exit 1 onto Hohensteinstraße", il.get(8).getTurnDescription(tr));
         //..
-        assertEquals("turn right onto Treppelweg", il.get(15).getTurnDescription(tr));
+        assertEquals("turn right onto Treppelweg", il.get(21).getTurnDescription(tr));
 
         rsp = hopper.route(new GHRequest(48.410987, 15.599492, 48.411172, 15.600371).
                 setAlgorithm(ASTAR).setProfile(footProfile));
@@ -1417,8 +1412,8 @@ public class GraphHopperTest {
                 .setProfile(bikeProfile));
         res = rsp.getBest();
         assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
-        assertEquals(536, res.getTime() / 1000f, 1);
-        assertEquals(2522, res.getDistance(), 1);
+        assertEquals(500, res.getTime() / 1000f, 1);
+        assertEquals(2211, res.getDistance(), 1);
 
         rsp = hopper.route(new GHRequest(43.73005, 7.415707, 43.741522, 7.42826)
                 .setProfile("profile3"));
@@ -2076,34 +2071,6 @@ public class GraphHopperTest {
         rsp = hopper.route(req);
         assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
         assertEquals(658, rsp.getBest().getDistance(), 1);
-    }
-
-    @Test
-    public void testTagParserProcessingOrder() {
-        // it does not matter when the OSMBikeNetworkTagParser is added (before or even after BikeCommonPriorityParser)
-        // as it is a different type but it is important that OSMSmoothnessParser is added before smoothnessEnc is used
-        // in BikeCommonAverageSpeedParser
-        GraphHopper hopper = new GraphHopper().
-                setGraphHopperLocation(GH_LOCATION).
-                setOSMFile(BAYREUTH).
-                setMinNetworkSize(0).
-                setEncodedValuesString("bike_access, bike_priority, bike_average_speed").
-                setProfiles(TestProfiles.accessSpeedAndPriority("bike"));
-
-        hopper.importOrLoad();
-        GHRequest req = new GHRequest(new GHPoint(49.98021, 11.50730), new GHPoint(49.98026, 11.50795));
-        req.setProfile("bike");
-        GHResponse rsp = hopper.route(req);
-        assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
-        // due to smoothness=bad => 7 seconds longer
-        assertEquals(21, rsp.getBest().getTime() / 1000.0, 1);
-
-        req = new GHRequest(new GHPoint(50.015067, 11.502093), new GHPoint(50.014694, 11.499748));
-        req.setProfile("bike");
-        rsp = hopper.route(req);
-        assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
-        // due to bike network (relation 2247905) a lower route weight => otherwise 29.0
-        assertEquals(23.2, rsp.getBest().getRouteWeight(), .1);
     }
 
     @Test