+++ core/src/test/java/com/graphhopper/routing/PathTest.java
@@ -18,10 +18,12 @@
 package com.graphhopper.routing;
 
 import com.carrotsearch.hppc.IntArrayList;
+import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.*;
 import com.graphhopper.routing.util.AllEdgesIterator;
 import com.graphhopper.routing.util.EncodingManager;
 import com.graphhopper.routing.util.TraversalMode;
+import com.graphhopper.routing.util.parsers.OrientationCalculator;
 import com.graphhopper.routing.weighting.SpeedWeighting;
 import com.graphhopper.routing.weighting.Weighting;
 import com.graphhopper.storage.BaseGraph;
@@ -46,8 +48,9 @@ import static org.junit.jupiter.api.Assertions.*;
 public class PathTest {
     private final DecimalEncodedValue carAvSpeedEnc = new DecimalEncodedValueImpl("speed", 5, 5, true);
     private final EncodingManager carManager = EncodingManager.start().add(carAvSpeedEnc).
-            add(VehicleAccess.create("car")).add(Roundabout.create()).add(RoadClass.create()).
-            add(RoadClassLink.create()).add(MaxSpeed.create()).build();
+            add(Orientation.create()).add(VehicleAccess.create("car")).add(Roundabout.create()).
+            add(RoadClass.create()).add(RoadEnvironment.create()).add(RoadClassLink.create()).
+            add(MaxSpeed.create()).build();
 
     private final DecimalEncodedValue mixedCarSpeedEnc = new DecimalEncodedValueImpl("mixed_car_speed", 5, 5, true);
     private final BooleanEncodedValue mixedCarAccessEnc = VehicleAccess.create("car");
@@ -56,6 +59,7 @@ public class PathTest {
             add(mixedCarAccessEnc).
             add(mixedCarSpeedEnc).add(mixedFootSpeedEnc).
             add(RoadClass.create()).
+            add(RoadEnvironment.create()).
             add(RoadClassLink.create()).
             add(MaxSpeed.create()).
             add(Roundabout.create()).build();
@@ -232,7 +236,7 @@ public class PathTest {
         // Test instructions
         List<String> tmpList = getTurnDescriptions(wayList);
         assertEquals(List.of("continue onto MainStreet 1 2",
-                        "At roundabout, take exit 3 onto 5-8",
+                        "at roundabout, take exit 3 onto 5-8",
                         "arrive at destination"),
                 tmpList);
         // Test Radian
@@ -246,7 +250,7 @@ public class PathTest {
         wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, mixedEncodingManager, tr);
         tmpList = getTurnDescriptions(wayList);
         assertEquals(List.of("continue onto MainStreet 1 2",
-                        "At roundabout, take exit 2 onto MainStreet 4 7",
+                        "at roundabout, take exit 2 onto MainStreet 4 7",
                         "arrive at destination"),
                 tmpList);
         // Test Radian
@@ -263,7 +267,7 @@ public class PathTest {
         assertTrue(p.isFound());
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, mixedEncodingManager, tr);
         List<String> tmpList = getTurnDescriptions(wayList);
-        assertEquals(List.of("At roundabout, take exit 3 onto 5-8",
+        assertEquals(List.of("at roundabout, take exit 3 onto 5-8",
                         "arrive at destination"),
                 tmpList);
     }
@@ -278,7 +282,7 @@ public class PathTest {
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, mixedEncodingManager, tr);
         List<String> tmpList = getTurnDescriptions(wayList);
         assertEquals(List.of("continue onto 3-6",
-                        "At roundabout, take exit 3 onto 5-8",
+                        "at roundabout, take exit 3 onto 5-8",
                         "arrive at destination"),
                 tmpList);
         roundaboutGraph.inverse3to9();
@@ -487,6 +491,34 @@ public class PathTest {
         assertEquals(intersectionMap, intersectionDetails.get(1).getValue());
     }
 
+    @Test
+    public void testChangeDetails() {
+        Weighting weighting = new SpeedWeighting(carAvSpeedEnc);
+        DecimalEncodedValue orEnc = carManager.getDecimalEncodedValue(Orientation.KEY);
+        OrientationCalculator calc = new OrientationCalculator(orEnc);
+        AllEdgesIterator allEdges = pathDetailGraph.getAllEdges();
+        EdgeIntAccess intAccess = pathDetailGraph.getBaseGraph().getEdgeAccess();
+        while(allEdges.next()) {
+            ReaderWay way = new ReaderWay(allEdges.getEdge());
+            way.setTag("point_list", allEdges.fetchWayGeometry(FetchMode.ALL));
+            calc.handleWayTags(allEdges.getEdge(), intAccess, way, null);
+        }
+        Path path = new Dijkstra(pathDetailGraph, weighting, TraversalMode.NODE_BASED).calcPath(1, 7);
+        assertTrue(path.isFound());
+
+        Map<String, List<PathDetail>> details = PathDetailsFromEdges.calcDetails(path, carManager, weighting,
+                List.of(CHANGE_ANGLE), new PathDetailsBuilderFactory(), 0, pathDetailGraph);
+        assertEquals(1, details.size());
+
+        List<PathDetail> caDetail = details.get(CHANGE_ANGLE);
+
+        assertEquals(4, caDetail.size());
+        assertNull(caDetail.get(0).getValue());
+        assertEquals(0.0, caDetail.get(1).getValue());
+        assertEquals(-132.0, caDetail.get(2).getValue());
+        assertEquals(72.0, caDetail.get(3).getValue());
+    }
+
     /**
      * case with one edge being not an exit
      */
@@ -500,7 +532,7 @@ public class PathTest {
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, mixedEncodingManager, tr);
         List<String> tmpList = getTurnDescriptions(wayList);
         assertEquals(List.of("continue onto MainStreet 1 2",
-                        "At roundabout, take exit 2 onto 5-8",
+                        "at roundabout, take exit 2 onto 5-8",
                         "arrive at destination"),
                 tmpList);
         // Test Radian
@@ -577,7 +609,7 @@ public class PathTest {
         assertTrue(p.isFound());
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, carManager, tr);
         List<String> tmpList = getTurnDescriptions(wayList);
-        assertEquals(List.of("At roundabout, take exit 1 onto MainStreet 1 11",
+        assertEquals(List.of("at roundabout, take exit 1 onto MainStreet 1 11",
                         "arrive at destination"),
                 tmpList);
     }
@@ -592,7 +624,7 @@ public class PathTest {
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, mixedEncodingManager, tr);
         List<String> tmpList = getTurnDescriptions(wayList);
         assertEquals(List.of("continue onto MainStreet 1 2",
-                        "At roundabout, take exit 1 onto 5-8",
+                        "at roundabout, take exit 1 onto 5-8",
                         "arrive at destination"),
                 tmpList);
         // Test Radian
@@ -689,6 +721,40 @@ public class PathTest {
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, carManager, tr);
 
         assertEquals(2, wayList.size());
+    }
+
+    @Test
+    public void testFerry() {
+        final BaseGraph graph = new BaseGraph.Builder(carManager).create();
+        final NodeAccess na = graph.getNodeAccess();
+
+        //      1 ---- 2 ---- 3 ---- 4
+        na.setNode(1, 48.909071, 8.647136);
+        na.setNode(2, 48.909071, 8.647978);
+        na.setNode(3, 48.909071, 8.648155);
+        na.setNode(3, 48.909071, 8.648200);
+
+        EnumEncodedValue<RoadEnvironment> roadEnvEnc = carManager.getEnumEncodedValue(RoadEnvironment.KEY, RoadEnvironment.class);
+
+        graph.edge(1, 2).set(carAvSpeedEnc, 60, 60).setDistance(5).set(roadEnvEnc, RoadEnvironment.ROAD).
+                setKeyValues(Map.of(STREET_NAME, new KValue("A B")));
+        graph.edge(2, 3).set(carAvSpeedEnc, 60, 60).setDistance(5).set(roadEnvEnc, RoadEnvironment.FERRY).
+                setKeyValues(Map.of(STREET_NAME, new KValue("B C")));
+        graph.edge(3, 4).set(carAvSpeedEnc, 60, 60).setDistance(5).set(roadEnvEnc, RoadEnvironment.ROAD).
+                setKeyValues(Map.of(STREET_NAME, new KValue("C D")));
+
+        Weighting weighting = new SpeedWeighting(carAvSpeedEnc);
+        Path p = new Dijkstra(graph, weighting, TraversalMode.NODE_BASED)
+                .calcPath(1, 4);
+        assertTrue(p.isFound());
+        InstructionList wayList = InstructionsFromEdges.calcInstructions(p, p.graph, weighting, carManager, tr);
+        assertEquals(4, wayList.size());
+        assertEquals("continue onto A B", wayList.get(0).getTurnDescription(tr));
+        assertEquals("Attention, take ferry (B C)", wayList.get(1).getTurnDescription(tr));
+        assertEquals(Instruction.FERRY, wayList.get(1).getSign());
+        assertEquals("leave ferry and turn right onto C D", wayList.get(2).getTurnDescription(tr));
+        assertEquals(Instruction.TURN_RIGHT, wayList.get(2).getSign());
+        assertEquals("arrive at destination", wayList.get(3).getTurnDescription(tr));
     }
 
     @Test
@@ -1035,6 +1101,7 @@ public class PathTest {
                 add(footAccessEnc).
                 add(RoadClass.create()).
                 add(RoadClassLink.create()).
+                add(RoadEnvironment.create()).
                 add(MaxSpeed.create()).
                 add(rdEnc).build();
 
@@ -1050,8 +1117,8 @@ public class PathTest {
         //      \     /
         //    7--8-->9<--10
 
-        na.setNode(0, 52.503809,13.410198);
-        na.setNode(1, 52.503871,13.410249);
+        na.setNode(0, 52.503809, 13.410198);
+        na.setNode(1, 52.503871, 13.410249);
         na.setNode(2, 52.503751, 13.410377);
         na.setNode(3, 52.50387, 13.410807);
         na.setNode(4, 52.503989, 13.41094);
@@ -1080,22 +1147,22 @@ public class PathTest {
         Path p = new Dijkstra(g, weighting, TraversalMode.NODE_BASED).calcPath(7, 10);
         assertEquals("[7, 8, 9, 10]", p.calcNodes().toString());
         InstructionList wayList = InstructionsFromEdges.calcInstructions(p, g, weighting, manager, tr);
-        assertEquals("At roundabout, take exit 1 onto Southeast in", wayList.get(1).getTurnDescription(tr));
+        assertEquals("at roundabout, take exit 1 onto Southeast in", wayList.get(1).getTurnDescription(tr));
 
         p = new Dijkstra(g, weighting, TraversalMode.NODE_BASED).calcPath(10, 1);
         assertEquals("[10, 9, 5, 3, 2, 1]", p.calcNodes().toString());
         wayList = InstructionsFromEdges.calcInstructions(p, g, weighting, manager, tr);
-        assertEquals("At roundabout, take exit 2 onto Nordwest, foot-only", wayList.get(1).getTurnDescription(tr));
+        assertEquals("at roundabout, take exit 2 onto Nordwest, foot-only", wayList.get(1).getTurnDescription(tr));
 
         p = new Dijkstra(g, weighting, TraversalMode.NODE_BASED).calcPath(10, 4);
         assertEquals("[10, 9, 5, 3, 4]", p.calcNodes().toString());
         wayList = InstructionsFromEdges.calcInstructions(p, g, weighting, manager, tr);
-        assertEquals("At roundabout, take exit 1 onto Nordeast in", wayList.get(1).getTurnDescription(tr));
+        assertEquals("at roundabout, take exit 1 onto Nordeast in", wayList.get(1).getTurnDescription(tr));
 
         p = new Dijkstra(g, weighting, TraversalMode.NODE_BASED).calcPath(10, 6);
         assertEquals("[10, 9, 5, 6]", p.calcNodes().toString());
         wayList = InstructionsFromEdges.calcInstructions(p, g, weighting, manager, tr);
-        assertEquals("At roundabout, take exit 1 onto Nordeast out", wayList.get(1).getTurnDescription(tr));
+        assertEquals("at roundabout, take exit 1 onto Nordeast out", wayList.get(1).getTurnDescription(tr));
     }
 
     static class AccessWeighting implements Weighting {
@@ -1153,18 +1220,24 @@ public class PathTest {
         final BaseGraph graph = new BaseGraph.Builder(carManager).create();
         final NodeAccess na = graph.getNodeAccess();
 
+        // 6 -5 \   / 7
+        //        4
+        //           \
+        //    1 - 2 - 3
         na.setNode(1, 52.514, 13.348);
         na.setNode(2, 52.514, 13.349);
         na.setNode(3, 52.514, 13.350);
         na.setNode(4, 52.515, 13.349);
         na.setNode(5, 52.516, 13.3452);
         na.setNode(6, 52.516, 13.344);
+        na.setNode(7, 52.516, 13.350);
 
         graph.edge(1, 2).set(carAvSpeedEnc, 45, 45).setDistance(5).setKeyValues(Map.of(STREET_NAME, new KValue("1-2")));
         graph.edge(4, 5).set(carAvSpeedEnc, 45, 45).setDistance(5).setKeyValues(Map.of(STREET_NAME, new KValue("4-5")));
         graph.edge(2, 3).set(carAvSpeedEnc, 90, 90).setDistance(5).setKeyValues(Map.of(STREET_NAME, new KValue("2-3")));
         graph.edge(3, 4).set(carAvSpeedEnc, 9, 9).setDistance(10).setKeyValues(Map.of(STREET_NAME, new KValue("3-4")));
         graph.edge(5, 6).set(carAvSpeedEnc, 9, 9).setDistance(0.001).setKeyValues(Map.of(STREET_NAME, new KValue("3-4")));
+        graph.edge(4, 7).set(carAvSpeedEnc, 45, 45).setDistance(5).setKeyValues(Map.of(STREET_NAME, new KValue("4-7")));
         return graph;
     }
 
@@ -1286,5 +1359,4 @@ public class PathTest {
     private static Path extractPath(Graph graph, Weighting weighting, SPTEntry sptEntry) {
         return PathExtractor.extractPath(graph, weighting, sptEntry);
     }
-
 }