+++ web-bundle/src/main/resources/com/graphhopper/maps/config.js
@@ -26,4 +26,5 @@ const config = {
         ],
         snapPreventions: ['ferry'],
     },
+    profile_group_mapping: {},
 }